Generate better names for array items
