+++ statham/__main__.py
@@ -1,9 +1,8 @@
 from argparse import ArgumentParser, RawTextHelpFormatter
-from collections import defaultdict
 from contextlib import contextmanager
 from logging import getLogger, INFO
 from os import path
-from typing import Any, Callable, DefaultDict, Dict, Iterator, TextIO, Tuple
+from typing import Any, Dict, Iterator, TextIO, Tuple
 from sys import argv, stdout
 
 from json_ref_dict import materialize, RefDict
@@ -12,6 +11,7 @@ from statham.constants import IGNORED_SCHEMA_KEYWORDS
 from statham.dependency_resolver import ClassDependencyResolver
 from statham.models import parse_schema
 from statham.serializer import serialize_object_schemas
+from statham.title_generator import title_labeller
 
 
 LOGGER = getLogger(__name__)
@@ -96,42 +96,6 @@ def _convert_schema(schema_dict: Dict[str, Any]) -> str:
     )
 
 
-def _get_title_from_reference(reference: str) -> str:
-    """Convert JSONSchema references to title fields.
-
-    If the reference has a pointer, use the final segment, otherwise
-    use the final segment of the base uri stripping any content type
-    extension.
-
-    :param reference: The JSONPointer reference.
-    """
-    reference = reference.rstrip("/")
-    base, pointer = reference.split("#")
-    if not pointer:
-        return base.split("/")[-1].split(".")[0]
-    return pointer.split("/")[-1]
-
-
-def _get_title_labeller() -> Callable[[str], Tuple[str, str]]:
-    """Create a title labeller, enumerating repeated titles.
-
-    Used to assign meaningful names to schemas which have no specified
-    title.
-    """
-    counter: DefaultDict[str, Iterator[int]] = defaultdict(
-        lambda: iter(range(0, 1000))
-    )
-
-    def _get_title(reference: str) -> Tuple[str, str]:
-        name = _get_title_from_reference(reference)
-        count = next(counter[name])
-        if count:
-            name = f"{name}{count}"
-        return "title", name
-
-    return _get_title
-
-
 def main(input_uri: str) -> str:
     """Get the schema, and then return the generated python module.
 
@@ -148,7 +112,7 @@ def main(input_uri: str) -> str:
     schema = materialize(
         RefDict(input_uri),
         exclude_keys=IGNORED_SCHEMA_KEYWORDS,
-        context_labeller=_get_title_labeller(),
+        context_labeller=title_labeller(),
     )
     return _convert_schema(schema)
